requirements.txt
django
time
beautifulsoup4
datetime
python-docx
wget
aspose-words 